Break apart the tenant and subscription data.

+++ src/connection.move.cpp
@@ -6,6 +6,7 @@
 */
 
 
+#include <rask/subscriber.hpp>
 #include <rask/tenant.hpp>
 
 #include <fost/log>
@@ -32,19 +33,21 @@ void rask::move_out(rask::connection::in &packet) {
     logger
         ("tenant", tenant->name())
         ("name", name);
-    packet.socket->workers.high_latency.get_io_service().post(
-        [tenant, name = std::move(name), priority]() {
-            auto location = tenant->local_path() /
-                fostlib::coerce<boost::filesystem::path>(name);
-            tenant->remote_change(location, tenant::move_inode_out, priority);
-            auto removed = boost::filesystem::remove_all(location);
-            if ( removed ) {
-                fostlib::log::warning(c_fost_rask)
-                    ("", "Deleting files")
-                    ("tenant", tenant->name())
-                    ("root", location)
-                    ("count", removed);
-            }
-        });
+    if ( tenant->subscription ) {
+        packet.socket->workers.high_latency.get_io_service().post(
+            [tenant, name = std::move(name), priority]() {
+                auto location = tenant->subscription->local_path() /
+                    fostlib::coerce<boost::filesystem::path>(name);
+                tenant->subscription->remote_change(location, tenant::move_inode_out, priority);
+                auto removed = boost::filesystem::remove_all(location);
+                if ( removed ) {
+                    fostlib::log::warning(c_fost_rask)
+                        ("", "Deleting files")
+                        ("tenant", tenant->name())
+                        ("root", location)
+                        ("count", removed);
+                }
+            });
+    }
 }
